refactor: rating 잘못된 타입 double->int로 변경

+++ matgpt/src/main/java/com/ktc/matgpt/review/dto/ReviewResponse.java
@@ -236,7 +236,7 @@ public class ReviewResponse {
     @Getter
     public static class RecentReviewDTO {
         private Long id;
-        private double rating;
+        private int rating;
         private String content;
         private LocalDateTime createdAt;
         private String storeImage;
